fix: importing a file create an infinite number of layers
> fix #209

This bug could easily be fixed by submitting a PR to __Kepler.gl__.
We'll probably give it a shot sometimes but with no great expectations.

File upload related actions weren't scoped to a __Kepler.gl__ instance slice.
We had to internalize a big part of the `visState` and root reducer
to override the related state `updaters`.