feat(layer-color): keep layer and dataset color in sync
This partially fix #230 there are still edge cases persisting.
Notably in `TooltipConfig` where the dataset color is used.
This is fixable in the UI when we update a layer color because it also change the dataset color.

This could be totally fixed by using the same color as the dataset at layer creation time.