feat: implement tooltip and detailed view of a POI
Add a detailed, fullscreen, view of all POI information.
This view is accessible from a button in the tooltip of a POI.

This roughly implement rules described #249 but should be enough for testing the feature.

This also loosely implements @SeongHunChandra design,
but will require some more hard work to get it right.